# Fix excessive allocation in the resource metrics