<!-- Thanks for sending a pull request! -->

# Changes

Through a bit of cleanup work in Knative Serving, it got kinda apparent that our metric recording calls are kinda expensive. I triaged it down to the `resourceToKey` being quite expensive.

Discussing with @evankanderson, we came up with this quick fix to make the allocations of the functions much more reasonable without coming up with a different way of handling the keys here altogether.

## Benchmark results

```
benchmark                               old ns/op     new ns/op     delta
BenchmarkResourceToKey/0-labels-16      139           2.38          -98.29%
BenchmarkResourceToKey/1-labels-16      677           168           -75.18%
BenchmarkResourceToKey/5-labels-16      2344          802           -65.78%
BenchmarkResourceToKey/10-labels-16     4625          1645          -64.43%

benchmark                               old allocs     new allocs     delta
BenchmarkResourceToKey/0-labels-16      2              0              -100.00%
BenchmarkResourceToKey/1-labels-16      6              1              -83.33%
BenchmarkResourceToKey/5-labels-16      18             3              -83.33%
BenchmarkResourceToKey/10-labels-16     33             3              -90.91%

benchmark                               old bytes     new bytes     delta
BenchmarkResourceToKey/0-labels-16      40            0             -100.00%
BenchmarkResourceToKey/1-labels-16      128           32            -75.00%
BenchmarkResourceToKey/5-labels-16      432           192           -55.56%
BenchmarkResourceToKey/10-labels-16     800           320           -60.00%
```

/assign @evankanderson